Simplified page navigations using simple hyperlinks

+++ TowerDefense/Interfaces/Point.cs
@@ -2,10 +2,10 @@
 {
     public struct Point
     {
-        public readonly int X;
-        public readonly int Y;
+        public readonly double X;
+        public readonly double Y;
 
-        public Point(int x, int y)
+        public Point(double x, double y)
         {
             this.X = x;
             this.Y = y;
